# banking-api/docs/swagger.yaml
definitions:
  definitions.Account:
    properties:
      account_number:
        type: string
      balance:
        type: number
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  definitions.Transaction:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      created_at:
        type: string
      currency_code:
        type: string
      id:
        type: integer
      status:
        type: string
      transaction_fee:
        type: number
      type:
        type: string
      updated_at:
        type: string
    type: object
  definitions.TransactionInput:
    properties:
      amount:
        type: number
      currency_code:
        enum:
        - USD
        - EUR
        type: string
      type:
        enum:
        - debit
        - credit
        type: string
    required:
    - amount
    type: object
  definitions.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  definitions.UserInput:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is a API representing a simple banking system.
  title: Plutus Banking API
  version: "1.0"
paths:
  /account/{id}/transaction:
    post:
      description: Create an account transaction in the system
      parameters:
      - description: transaction to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/definitions.TransactionInput'
      - description: id of account to create transaction in
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/definitions.Transaction'
      summary: Create an account transaction
      tags:
      - account
  /healthz:
    get:
      description: Check service health condition
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Check service health
      tags:
      - health
  /user:
    post:
      description: Create a user in the system
      parameters:
      - description: user to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/definitions.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/definitions.User'
      summary: Create a user
      tags:
      - user
  /user/{id}:
    delete:
      description: Delete a user from the system
      parameters:
      - description: id of user to delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Delete a user
      tags:
      - user
    get:
      description: Retrieve a user in the system
      parameters:
      - description: id of user to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/definitions.User'
      summary: Retrieve a user
      tags:
      - user
    put:
      description: Update a user in the system
      parameters:
      - description: id of user to update
        in: path
        name: id
        required: true
        type: string
      - description: user to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/definitions.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/definitions.User'
      summary: Update a user
      tags:
      - user
  /user/{id}/account:
    post:
      description: Create a user account
      parameters:
      - description: id of user for whom to create the account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/definitions.Account'
      summary: Create a user account
      tags:
      - account
  /user/{user_id}/account/{account_id}:
    get:
      description: Retrieve a user account in the system
      parameters:
      - description: id of user to retrieve
        in: path
        name: user_id
        required: true
        type: string
      - description: id of account to retrieve
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/definitions.Account'
      summary: Retrieve a user account
      tags:
      - account
  /users:
    get:
      description: Retrieve a list of users in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/definitions.User'
              type: array
            type: array
      summary: Retrieve a list of users
      tags:
      - user
swagger: "2.0"